+++ src/policy_manager/UniformGroup.java
@@ -1,9 +1,9 @@
 package policy_manager;
 import java.util.List;
 
-import RMI.InvokerInterface;
-import application.Invoker;
 import faas_exceptions.NoInvokerAvailable;
+import invoker.Invoker;
+import invoker.InvokerInterface;
 
 /**
  * Implementation of the PolicyManager interface using a uniform distribution
@@ -37,7 +37,7 @@ public class UniformGroup implements PolicyManager{
      * @throws NoInvokerAvailable If no suitable Invoker can be found within the specified retries.
      */
     @Override
-    public InvokerInterface getInvoker(List<InvokerInterface> invokers, int ram) throws Exception {
+    public InvokerInterface getInvoker(List<InvokerInterface> invokers, long ram) throws Exception {
         InvokerInterface invoker = invokers.get(lastInvokerAssigned);
 
         if (invokers.isEmpty()) throw new NoInvokerAvailable("No Invokers in list.");
